requirements.txt
asgiref==3.9.1
Django==5.2.6
pillow==11.3.0
PyMySQL==1.1.2
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
xendit-python==7.0.0
